Read these uploaded files
These files are working examples of how we're going to be able to add the feature to post, to Twitter, and to LinkedIn. The reason why I added these files is so that you, the AI model, can see the working example and apply it to our app. Thank you.